Added better view structure and changed last part of contact and socials
