+++ src/views/index/CompleteExamples.js
