+++ src/views/index/EventExplanation.js
