+++ src/views/index/Examples.js
