+++ src/views/index/Tabs.js
